fix(coverage): avoid eval to allow coverage
The previous method of putting the content in a single function
call as a string prevented subsequent passes from the coverage
preprocessor from properly attaching to that code. This uses the
non-eval version of the call to goog.loadModule() to allow coverage
to function properly.

I kept the sourceURL line but I don't think it is doing anything.
However, I was able to debug in Chrome with this tiny wrapper script
without issues.